add split screen grid option
Before, I was trying to be smart about automatically using the
half-width grid when there were fewer datums in each chart. But after
using Charted for a while, the # datums isn’t a great indicator and the
desired grid is almost always half-width with a few exceptions. So I’m
switching that to the default and making an explicit option for
full-width charts.