Upgraded flow to 0.37.4
Flow is now an official NPM dependency for this package. This new
version wasn't happy with my ad-hoc interface for Express so, after
spending an hour or so battling it, I decided to install flow-typed
which has a very verbose interface for Express (hence all the changes
in server/charted and other files).

Also, Flow doesn't understand the non-standard `unescape` function
in JavaScript so I've decided to simply suppress that warning.

I also spent way too much time trying to type-check req.body and then
gave up and converted it to `any`. In future, it'd be nice to have
something like proto-bufs for request bodies.