Make title and note fields to preserve line breaks
There's a new class in town: Editor! It's not really an editor
but a very ad-hoc wrapper around content editable fields. It uses
`innerText` to get text as it is presented to the user [1] on
`getContent` and converts line breaks into DIVs on `setContent`.
It also alerts upstream about changes and sets/unsets the `empty`
CSS class.

[1] - Before we used jQuery's text() function to get text
      contents of an element. But jQuery uses textContent property
      which doesn't preserve line breaks. innerText suits better
      here but has its own quirks. Oh, web.

      More context on innerText vs. textContent:
        http://perfectionkills.com/the-poor-misunderstood-innerText/

Fixes #27